初步完善实现回复功能api

+++ back/src/main/java/com/su/community/controller/PublishController.java
@@ -43,7 +43,7 @@ public class PublishController {
     public String doPublish(String title,
                             String description,
                             String tag,
-                            Integer id,
+                            Long id,
                             HttpServletRequest request,
                             Model model) {
         model.addAttribute("title", title);
@@ -89,7 +89,7 @@ public class PublishController {
     }
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id")Integer id,Model model){
+    public String edit(@PathVariable(name = "id")Long id,Model model){
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title",question.getTitle());
         model.addAttribute("description", question.getDescription());
